Removes redundant test macro
Replaces and removes the 'ExpressibleByExprSyntaxMacro' test macro in
favor of the 'Macro0' test macro.
